server: add llama_server_queue struct
